# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SecureAI - Biometric Security Demo API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /security/logs:
    get:
      operationId: listSecurityLogs
      tags: [security]
      summary: List security audit logs
      parameters:
        - in: query
          name: limit
          schema:
            type: number
            default: 50
        - in: query
          name: offset
          schema:
            type: number
            default: 0
        - in: query
          name: eventType
          schema:
            type: string
        - in: query
          name: userEmail
          description: Case-insensitive partial match against the actor's email
          schema:
            type: string
        - in: query
          name: ipAddress
          description: Case-insensitive partial match against the request's origin IP
          schema:
            type: string
        - in: query
          name: fromDate
          description: ISO 8601 timestamp — only logs at or after this time
          schema:
            type: string
        - in: query
          name: toDate
          description: ISO 8601 timestamp — only logs at or before this time
          schema:
            type: string
      responses:
        "200":
          description: Security logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityLog"
...
  /security/logs/verify:
    get:
      operationId: verifyLogIntegrity
      tags: [security]
      summary: Recompute the audit log hash chain and report whether it's intact (admin only)
      responses:
        "200":
          description: Chain verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogChainVerification"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /security/logs/repair:
    post:
      operationId: repairLogChain
      tags: [security]
      summary: Quarantine the untrustworthy tail of a broken hash chain and log the repair (security_analyst only)
      responses:
        "200":
          description: Repair result — no-op if the chain was already valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChainRepairResult"
        "403":
          description: Security analyst access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /security/logs/restore:
    post:
      operationId: restoreLogChain
      tags: [security]
      summary: Re-insert deleted rows from the deletion-audit trigger's pre-delete snapshots, and log the restoration (security_analyst only)
      responses:
        "200":
          description: Restore result — no-op if the chain was already valid, or if the break wasn't caused by a deletion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChainRestoreResult"
        "403":
          description: Security analyst access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /security/logs/deletions:
    get:
      operationId: listDeletionAudit
      tags: [security]
      summary: Forensic view of every security_logs deletion the database trigger has captured, app-initiated or raw SQL alike (security_analyst only)
      responses:
        "200":
          description: Deletion audit entries, most recent first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeletionAuditEntry"
        "403":
          description: Security analyst access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SecurityLog:
      type: object
      required: [id, eventType, timestamp, details]
      properties:
        id:
          type: number
        userId:
          type: ["number", "null"]
        userEmail:
          type: ["string", "null"]
        eventType:
          type: string
        ipAddress:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        details:
          type: string
        timestamp:
          type: string
...
    LogChainVerification:
      type: object
      required: [valid, rowsChecked, brokenAtId, reason]
      properties:
        valid:
          type: boolean
        rowsChecked:
          type: number
        brokenAtId:
          type: ["number", "null"]
        reason:
          type: ["string", "null"]
    ChainRepairResult:
      type: object
      required: [repaired, removedCount, removedFromId, verification]
      properties:
        repaired:
          description: False if the chain was already valid — nothing was removed
          type: boolean
        removedCount:
          type: number
        removedFromId:
          type: ["number", "null"]
        verification:
          description: Re-verification result after the repair (or the original check, if nothing needed repairing)
          $ref: "#/components/schemas/LogChainVerification"
    ChainRestoreResult:
      type: object
      required: [restored, restoredCount, restoredIds, unrecoverableIds, verification]
      properties:
        restored:
          type: boolean
        restoredCount:
          type: number
        restoredIds:
          type: array
          items:
            type: number
        unrecoverableIds:
          description: Ids confirmed missing but with no deletion-audit snapshot to restore from
          type: array
          items:
            type: number
        verification:
          description: Re-verification result after restoring (or the original check, if nothing needed restoring)
          $ref: "#/components/schemas/LogChainVerification"
    DeletionAuditEntry:
      type: object
      required: [id, deletedLogId, eventType, details, deletedByAppActor, deletedByDbRole, deletedByClientAddr, deletedAt, currentlyRestored]
      properties:
        id:
          type: number
        deletedLogId:
          description: The security_logs id that was deleted
          type: number
        eventType:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        deletedByAppActor:
          description: Set only for app-initiated deletions (e.g. repairLogChain) — null means raw/out-of-band, such as a direct SQL client
          type: ["string", "null"]
        deletedByDbRole:
          type: string
        deletedByClientAddr:
          type: ["string", "null"]
        deletedAt:
          type: string
        currentlyRestored:
          description: Whether a row now sits at this id again (e.g. restored via restoreLogChain)
          type: boolean
...